add sharing and task-to-kill agreement

+++ src/main/java/ru/sbtqa/tag/pagefactory/stepdefs/SetupStepDefs.java
@@ -53,7 +53,7 @@ public class SetupStepDefs {
 
         try {
             String tasksToKill = Props.get("tasks.to.kill");
-            if (!"".equals(tasksToKill)) {
+            if (!PageFactory.isSharingProcessing() && !"".equals(tasksToKill)) {
                 for (String task : tasksToKill.split(",")) {
                     if (SystemUtils.IS_OS_WINDOWS) {
                         Runtime.getRuntime().exec("taskkill /IM " + task.trim() + " /F");
@@ -123,6 +123,7 @@ public class SetupStepDefs {
 
         if (PageFactory.getEnvironment() == Environment.WEB && TagWebDriver.isWebDriverShared()) {
             LOG.info("Webdriver sharing is processing...");
+            PageFactory.setIsSharingProcessing(true);
         } else {
             PageFactory.dispose();
         }
